Fake/requirements.txt
# streamlit
# openai
# requests
# python-dotenv
# langchain

streamlit
fastapi
uvicorn
openai>=1.0.0
python-dotenv
requests